bbsBayes2: Hierarchical Bayesian Analysis of North American BBS Data

The North American Breeding Bird Survey (BBS) is a long-running program that seeks to monitor the status and trends of the breeding birds in North America. Since its start in 1966, the BBS has accumulated over 50 years of data for over 500 species of North American Birds. Given the temporal and spatial structure of the data, hierarchical Bayesian models are used to assess the status and trends of these 500+ species of birds. 'bbsBayes2' allows you to perform hierarchical Bayesian analysis of BBS data. You can run a full model analysis for one or more species that you choose, or you can take more control and specify how the data should be stratified, prepared for 'JAGS', or modelled. The functions provided here allow you to replicate analyses performed by the United State Geological Survey (USGS, see Link and Sauer (2011) doi:10.1525/auk.2010.09220 ) and Canadian Wildlife Service (CWS, see Smith and Edwards (2020) doi:10.1101/2020.03.26.010215 ).

See also

Author

Maintainer: Brandon P.M. Edwards brandonedwards3@cmail.carleton.ca

Authors:

Adam C. Smith adam.smith2@canada.ca

Steffi LaZerte steffi@steffi.ca