Common arguments and documentation for various functions

Common arguments and documentation for various functions

Arguments

level

Character. Which type of BBS counts to use, "state" or "stop". Default "state".

release

Numeric. Which yearly release to use, 2022 (including data through 2021 field season) or 2020 (including data through 2019). Default 2022.

sf

Spatial data frame. PLACEHOLDER

model

Character. Type of model to use, must be one of "first_diff" (First Differences), "gam" (General Additive Model), "gamye" (General Additive Model with Year Effect), or "slope" (Slope model).

model_variant

Character. Model variant to use, must be one of "nonhier" (Non-hierarchical), "hier" (Hierarchical; default), or "spatial" (Spatially explicit).

min_year

Numeric. Minimum year to use. Default (NULL) uses first year in data.

max_year

Numeric. Maximum year to use. Default (NULL) uses first year in data.

strata_data

List. Stratified data generated by stratify()

prepared_data

List. Prepared data generated by prepare_data().

spatial_data

List. Spatial neighbourhoods generated by prepare_spatial().

model_data

List. Model data generated by prepare_model().

model_output

List. Model output generated by run_model().

indices

List. Indices generated by generate_indices().

trends

List. Trends generated by generate_trends().

ci_width

Numeric. Quantile defining the width of the plotted credible interval. Defaults to 0.95 (lower = 0.025 and upper = 0.975). Note these quantiles need to have been precalculated in generate_indices().

add_observed_means

Logical. Whether to include points indicating the observed mean counts. Default FALSE. Note: scale of observed means and annual indices may not match due to imbalanced sampling among routes.

col_viridis

Logical. Should the colour-blind-friendly "viridis" palette be used. Default FALSE.

variables

Character vector. Specific variables (e.g., "strata_raw[1]") or variable types (e.g., "strata_raw") for which to calculate metrics. If NULL (default) all variables are returned.

set_seed

Numeric. If NULL (default) no seed is set. Otherwise an integer number to be used with withr::with_seed() internally to ensure reproducibility.

quiet

Logical. Suppress progress messages? Default FALSE.

Details

Use @inheritParams common_docs to include the above in any function documentation with a matching argument (will only include matching args)

NatureCounts account

All public data is available with a username/password (sign up for a free NatureCounts account). However, to access private/semi-public projects/collections you must request access. See the Access and request_ids section for more information.