Fetch Breeding Bird Survey dataset

Fetch and download Breeding Bird Survey data from the United States Geological Survey (USGS) FTP site. This is the raw data that is uploaded to the site before any analyses are performed. Users can download different types (state, stop) and different releases (currently 2020 and 2022).

Usage

fetch_bbs_data(
  level = "state",
  release = 2022,
  force = FALSE,
  quiet = FALSE,
  compression = "none"
)

Arguments

level

Character. Which type of BBS counts to use, "state" or "stop". Default "state".

release

Numeric. Which yearly release to use, 2022 (including data through 2021 field season) or 2020 (including data through 2019). Default 2022.

force

Logical. Should pre-exising BBS data be overwritten? Default FALSE.

quiet

Logical. Suppress progress messages? Default FALSE.

compression

Character. What compression should be used to save data? Default is none which takes up the most space but is the fastest to load. Must be one of none, gz, bz2, or xz (passed to readr::write_rds()'s compress argument).

Details

Users will be asked before saving the BBS data to a package-specific directory created on their computer. Before downloading any data, users must thoroughly read through the USGS terms and conditions for that data and enter the word "yes" to agree.

BBS state level counts provide counts beginning in 1966, aggregated in five bins, each of which contains cumulative counts from 10 of the 50 stops along a route. In contrast BBS stop level counts provides stop-level data beginning in 1997, which includes counts for each stop along routes individually. Note that stop-level data is not currently supported by the modelling utilities in bbsBayes2.

There are two releases for each type of data, 2020 and 2022. By default all functions use the most recent release unless otherwise specified. For example, the release argument in stratify() can be changed to 2020 to use the 2020 release of state-level counts.

See also

Examples

if (FALSE) { # interactive()

fetch_bbs_data(force = TRUE)
fetch_bbs_data(level = "stop", force = TRUE)
fetch_bbs_data(release = 2020, force = TRUE)
fetch_bbs_data(release = 2020, level = "stop", force = TRUE)
}