Generate a map of trends by strata

plot_map() allows you to generate a colour-coded map of the percent change in species trends for each strata.

Usage

plot_map(
  trends,
  slope = FALSE,
  title = TRUE,
  alternate_column = NULL,
  col_viridis = FALSE,
  strata_custom = NULL
)

Arguments

trends

List. Trends generated by generate_trends().

slope

Logical. Whether or not to map values of the alternative trend metric (slope of a log-linear regression) if slope = TRUE was used in generate_trends(), through the annual indices. Default FALSE.

title

Logical. Whether or not to include a title with species. Default TRUE.

alternate_column

Character, Optional name of numerical column in trends dataframe to plot. If one of the columns with "trend" in the title, (e.g., trend_q_0.05 then the colour scheme and breaks will match those used in the default trend maps)

col_viridis

Logical. Should the colour-blind-friendly "viridis" palette be used. Default FALSE.

strata_custom

(sf) Data Frame. Data frame of modified existing stratification, or a sf spatial data frame with polygons defining the custom stratifications. See details on strata_custom in stratify().

Value

a ggplot2 plot

See also

Examples

# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices
i <- generate_indices(pacific_wren_model)
#> Error in assert_valid_draws_format(format): could not find function "assert_valid_draws_format"

# Now generate trends
t <- generate_trends(i, slope = TRUE)
#> Error in eval(expr, envir, enclos): object 'i' not found

# Generate the map (without slope trends)
plot_map(t)
#> Error: `trends` must a list created by `generate_trends()` containing
#> `meta_data`, `meta_strata`, `trends`, `raw_data`

# Generate the map (with slope trends)
plot_map(t, slope = TRUE)
#> Error: `trends` must a list created by `generate_trends()` containing
#> `meta_data`, `meta_strata`, `trends`, `raw_data`

# Viridis
plot_map(t, col_viridis = TRUE)
#> Error: `trends` must a list created by `generate_trends()` containing
#> `meta_data`, `meta_strata`, `trends`, `raw_data`

# Generate a map (with alternate column - lower 95% Credible limit)
plot_map(t, alternate_column = "trend_q_0.05")
#> Error: `trends` must a list created by `generate_trends()` containing
#> `meta_data`, `meta_strata`, `trends`, `raw_data`