bbsBayes2: Hierarchical Bayesian Analysis of North American BBS Data

The North American Breeding Bird Survey (BBS) is a long-running program that seeks to monitor the status and trends of the breeding birds in North America. Since its start in 1966, the BBS has accumulated over 50 years of data for over 500 species of North American Birds. Given the temporal and spatial structure of the data, hierarchical Bayesian models are used to assess the status and trends of these 500+ species of birds. 'bbsBayes2' allows you to perform hierarchical Bayesian analysis of BBS data. You can run a full model analysis for one or more species, or you can take greater more control and specify how the data should be stratified, prepared for 'Stan', or modelled.

See also

Author

Maintainer: Brandon P.M. Edwards brandonedwards3@cmail.carleton.ca

Authors:

Adam C. Smith adam.smith@ec.gc.ca

Steffi LaZerte steffi@steffi.ca