Sample BBS data

Contains only Pacific Wren data

Usage

bbs_data_sample

Format

bbs_data_sample

A list containing:

birds - counts of each bird seen per route per

routes - data for each route run per year

species - species list of North America

Source

https://www.sciencebase.gov/ via fetch_bbs_data()

Details

A sample dataset containing only data for Pacific Wrens for the 2022 state-level BBS data. The full count set is obtained via the function fetch_bbs_data(). The data is obtained from the United States Geological Survey and is subject to change as new data is added each year. See References for citation.

References

Ziolkowski Jr., D.J., Lutmerding, M., Aponte, V.I., and Hudson, M-A.R., 2022, North American Breeding Bird Survey Dataset 1966 - 2021: U.S. Geological Survey data release, https://doi.org/10.5066/P97WAZE5