Return the cmdstanr summary

Extract and return the model summary using cmdstanr::summary().

Usage

get_summary(model_output, variables = NULL)

Arguments

model_output

List. Model output generated by run_model().

variables

Character vector. Specific variables (e.g., "strata_raw[1]") or variable types (e.g., "strata_raw") for which to calculate metrics. If NULL (default) all variables are returned.

Value

A data frame of model summary statistics, matching the output of cmdstanr::summary() plus one additional column variable_type that identifies categories of parameters (e.g., n for all of the annual indices for every stratum and year).

See also

Examples

# Temporarily suppress convergence warning for legibility
# "The ESS has been capped to avoid unstable estimates."
opts <- options(warn = -1)

# Using the example model for Pacific Wrens

get_summary(pacific_wren_model)
#> Error in assert_valid_draws_format(format): could not find function "assert_valid_draws_format"
get_summary(pacific_wren_model, variables = "strata_raw")
#> Error in assert_valid_draws_format(format): could not find function "assert_valid_draws_format"
get_summary(pacific_wren_model, variables = "strata_raw[9]")
#> Error in assert_valid_draws_format(format): could not find function "assert_valid_draws_format"

# Restore warnings
options(opts)